Task #2115: Merge with trunk 06032012

+++ branches/PVSS310-Bug1719-task-branch/MAC/APL/APLCommon/include/APL/APLCommon/ChildControl.h
@@ -167,8 +167,8 @@ private:
 	map <string, GCFTCPPort*>	itsStartDaemonMap;	// map<hostname,sdconnection>
 	typedef map<string, GCFTCPPort*>::iterator			SDiter;
 	typedef map<string, GCFTCPPort*>::const_iterator	const_SDiter;
-	uint32						itsStartupRetryInterval;
-	uint32						itsMaxStartupRetries;
+	int32						itsStartupRetryInterval;
+	int32						itsMaxStartupRetries;
 
 	list<ControllerInfo>*		itsCntlrList;		// admin. of child controllers
 
@@ -178,7 +178,7 @@ private:
 	uint32						itsActionTimer;		// ID of actiontimer.
 
 	uint32						itsGarbageTimer;	// for garbage collection
-	uint16						itsGarbageInterval;	// seconds between garbage coll.
+	uint32						itsGarbageInterval;	// seconds between garbage coll.
 
 	GCFTimerPort*				itsCompletionTimer;	// to signal parent in situation B
 	GCFITCPort*					itsCompletionPort;	// to signal parent in situation C
